pyproject.toml
[project]
name = "j-hard-tts-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "gdown>=5.2.1",
    "jiwer>=4.0.0",
    "numpy==1.26.4",
    "omegaconf>=2.3.0",
    "pyopenjtalk>=0.4.1",
    "reazonspeech-espnet-asr",
    "soxr>=1.0.0",
    "torchcodec>=0.9.1",
    "transformers>=4.57.6",
]

[tool.uv.sources]
reazonspeech-espnet-asr = { git = "https://github.com/reazon-research/ReazonSpeech.git", subdirectory = "pkg/espnet-asr" }
